+++ lib/features/logs/domain/usecases/update_log_item_usecase.dart
@@ -12,7 +12,7 @@ class UpdateLogItemParams {
   final int carbsGrams;
   final int proteinGrams;
   final int fatsGrams;
-  final int? portions;
+  final double? portions;
   final int? healthScore;
   final String? imageUrl;
   final List<IngredientEntity>? ingredients;
